Mandatory part done

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: victode- <victode-@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/30 00:38:57 by victode-          #+#    #+#             */
-/*   Updated: 2025/11/02 00:33:47 by victode-         ###   ########.fr       */
+/*   Updated: 2025/11/06 01:54:53 by victode-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
